#方案去皮值可以为小数,方案细节如果为包,设置去皮值为null

+++ src/main/java/com/zero/egg/model/Specification.java
@@ -63,7 +63,7 @@ public class Specification {
      * 数值（去皮对应的数值）
      */
     @ApiModelProperty(value = "数值（去皮对应的数值）",required=false)
-    private Integer numerical;
+    private Double numerical;
 
     /**
      * 去皮时,拼接"实重("+重量区间+")"
