+++ src/main/java/com/zero/egg/requestDTO/SpecificationRequestDTO.java
@@ -67,7 +67,7 @@ public class SpecificationRequestDTO implements Serializable {
      * 数值（去皮对应的数值）
      */
     @ApiModelProperty(value = "数值（去皮对应的数值）",required=false)
-    private Integer numerical;
+    private Double numerical;
 
     /**
      * 是否预警(0:否,1:是)
